Improvement controllerPost and 'index + unanswered + votes' done

+++ controller/ControllerPost.php
@@ -6,42 +6,63 @@ require_once 'framework/Controller.php';
 
 //A voir s'il ne faut pas ajouter un controller Answer
 class ControllerPost extends Controller{
-    
+
     public function index() {
-        (new View("index"))->show(array("posts"=> Question::get_questions()));
+        $user = null;
+        if(self::get_user_or_false())
+            $user = self::get_user_or_redirect();
+        (new View("index"))->show(array("posts"=> Question::get_questions(), "user" => $user));
     }
 
-    public function ask(){
-        $title = '';
-        $body = '';
-        $errors = [];
-        if (self::user_logged() && isset($_POST['title']) && isset($_POST['body'])){
-            $title = $_POST['title'];
-            $body = $_POST['body'];
-            $question = new Question(null, $_SESSION['user']->getUserId(), $title , $body, null, null, null, null, null, null, null);
-            $errors = Question::validate($question);
-            if(count($errors) == 0){
-                $question->create_question();   
-                self::redirect();
-            }
-        }
-        (new View("ask"))->show(array("title"=>$title, "body"=> $body, "errors" => $errors)); 
+    public function unanswered() {
+        $user = null;
+        if(self::get_user_or_false())
+            $user = self::get_user_or_redirect();
+        (new View("unanswered"))->show(array("posts"=> Question::get_questions_unanswered(), "user" => $user));
+    }
+
+    public function votes(){
+        $user = null;
+        if(self::get_user_or_false())
+            $user = self::get_user_or_redirect();
+        (new View("votes"))->show(array("posts"=> Question::get_questions_by_votes(), "user" => $user));    
     }
 
     //sert à afficher la view_show
     public function show(){
+        $user = null;
+        if(self::get_user_or_false())
+            $user = self::get_user_or_redirect();
         if(isset($_GET['param1'])){       
             $postId = $_GET['param1'];
             $result = Question::get_question($postId);
             $error = [];
             if($result){
-                (new View("show"))->show(array("post"=> $result, "error" => $error));
+                (new View("show"))->show(array("post"=> $result, "error" => $error, "user" => $user));
             }else{
                 self::redirect();
             }
         }else{
             self::redirect();
         }        
+    }
+
+    public function ask(){
+        $title = '';
+        $body = '';
+        $errors = [];
+        $user = self::get_user_or_redirect();
+        if (isset($_POST['title']) && isset($_POST['body'])){
+            $title = $_POST['title'];
+            $body = $_POST['body'];
+            $question = new Question(null, $user->getUserId(), $title , $body, null, null, null, null, null, null, null);
+            $errors = Question::validate($question);
+            if(count($errors) == 0){
+                $question->create_question();   
+                self::redirect();
+            }
+        }
+        (new View("ask"))->show(array("title"=>$title, "body"=> $body, "errors" => $errors, "user" => $user)); 
     }
 
     /*
@@ -64,86 +85,87 @@ class ControllerPost extends Controller{
         self::redirect();    
     }
 
-    
     //Gère l'édition d'une post ou d'une réponse
     public function edit(){
         $questionId = '';
         $answerId = '';
         $error = [];
+        $user = self::get_user_or_redirect();
         if(!isset($_POST['body']) && isset($_GET['param1'])){
             $questionId = $_GET['param1'];
             if(isset($_GET['param2'])){
                 $answerId = $_GET['param2'];
-                self::view_answer_edition($questionId, $answerId, $error);
+                self::view_answer_edition($questionId, $answerId, $user, $error);
             }else{
-                self::view_question_edition($questionId, $answerId, $error);
+                self::view_question_edition($questionId, $answerId, $user, $error);
             }
         }else{
-            self::edition_post();
+            self::edition_post($user);
         }
     }
 
     //Affiche la vue pour l'édition d'une réponse
-    private function view_answer_edition($questionId, $answerId, $error){
+    private function view_answer_edition($questionId, $answerId, $user, $error){   
         if (Question::get_question($questionId) &&  Answer::get_answer($answerId)){
             if(!is_numeric($questionId) || !is_numeric($answerId) 
-                                        || !(self::get_user_or_false()->getUserId() === Answer::get_answer($answerId)->getAuthorId())){
+                                        || !($user->getUserId() === Answer::get_answer($answerId)->getAuthorId())){
                 self::redirect();        
             }
         }else{
             self::redirect();
         }
-        (new View("edit"))->show(array("parentId"=>$questionId, "answerId" => $answerId, "post" => Answer::get_answer($answerId), 'error' => $error));         
+        (new View("edit"))->show(array("parentId"=>$questionId, "answerId" => $answerId, 
+                                       "post" => Answer::get_answer($answerId),"user" => $user, "error" => $error));         
     }
 
     //Affiche la vue pour l'edition d'une question
-    private function view_question_edition($questionId, $answerId, $errors){
+    private function view_question_edition($questionId, $answerId, $user, $errors){
         if(!is_numeric($questionId)){
             self::redirect();
         }
         if (Question::get_question($questionId)){
-            if(!(self::get_user_or_false()->getUserId() === Question::get_question($questionId)->getAuthorId())){
+            if(!($user->getUserId() === Question::get_question($questionId)->getAuthorId())){
                 self::redirect();     
             }
         }else{
             self::redirect();
         }
-        (new View("edit"))->show(array("parentId"=>$questionId, "answerId" => $answerId, "post" => Question::get_question($questionId), "error" => $errors));         
+        (new View("edit"))->show(array("parentId"=>$questionId, "answerId" => $answerId, 
+                                       "post" => Question::get_question($questionId), "user" => $user, "error" => $errors));         
     }
-
     
     //Sert à ajouter la modification du corps d'une réponse ou d'une question (à modifier).
-    private function edition_post(){
+    private function edition_post($user){
         if (isset($_GET['param1']) && isset($_POST['body'])){
             $questionId = $_GET['param1'];
             $body = $_POST['body'];
             $answerId = '';
             if(isset($_GET['param2'])){
                 $answerId = $_GET['param2'];
-                self::answer_edition($questionId, $answerId, $body);
+                self::answer_edition($questionId, $answerId, $user, $body);
             }else{
-                self::question_edition($questionId, $answerId, $body);
+                self::question_edition($questionId, $answerId, $user, $body);
             }                                    
         }else{
             self::redirect();
         }
     }
     
-    private function answer_edition($parentId, $answerId, $body){
-        if(self::get_user_or_redirect()->getUserId() === Answer::get_answer($answerId)->getAuthorId()){
+    private function answer_edition($parentId, $answerId, $user, $body){
+        if($user->getUserId() === Answer::get_answer($answerId)->getAuthorId()){
             $answer = new Answer($body, null, $parentId, null, null, $answerId, null);
             $error = Answer::validate($answer);
             if(count($error) == 0){
                 $answer->set_post();
                 self::redirect("post", "show", $parentId);       
             }else{
-                self::view_answer_edition($parentId, $answerId, $error);
+                self::view_answer_edition($parentId, $answerId, $user, $error);
             }
         }   
     }
 
-    private function question_edition($questionId,  $answerId, $body){
-        if(self::get_user_or_redirect()->getUserId() === Question::get_question($questionId)->getAuthorId()){
+    private function question_edition($questionId,  $answerId, $user, $body){
+        if($user->getUserId() === Question::get_question($questionId)->getAuthorId()){
             $title = Question::get_question($questionId)->getTitle();
             $question = new Question($questionId, null, $title, $body, null, null, null, null, null, null, null);
             $errors = Question::validate($question);
@@ -151,7 +173,7 @@ class ControllerPost extends Controller{
                 $question->set_post();
                 self::redirect("post", "show", $questionId);       
             }else{
-                self::view_question_edition($questionId, $answerId, $errors);
+                self::view_question_edition($questionId, $answerId, $user, $errors);
             }
         }
  
@@ -161,6 +183,7 @@ class ControllerPost extends Controller{
     public function delete(){
         $questionId = '';
         $answerId = '';
+        $user = self::get_user_or_redirect();
         if(self::get_user_or_redirect() && isset($_GET['param1'])){
             $questionId = $_GET['param1'];
             if(isset($_POST['cancel'])){
@@ -170,9 +193,9 @@ class ControllerPost extends Controller{
                 $answerId = $_GET['param2'];
             }
             if(isset($_POST['delete'])){
-                self::show_delete_questionOrAnswer($questionId, $answerId);    
+                self::show_delete_questionOrAnswer($questionId, $answerId, $user);    
             }else if(isset($_POST['delete_confirmation'])){
-                self::delete_questionOrAnswer($questionId, $answerId);   
+                self::delete_questionOrAnswer($questionId, $answerId, $user);   
             }else{
                 self::redirect();
             } 
@@ -186,63 +209,62 @@ class ControllerPost extends Controller{
         self::redirect("post", "show", $postId);      
     }
 
-    private function show_delete_questionOrAnswer($questionId, $answerId){
+    private function show_delete_questionOrAnswer($questionId, $answerId, $user){
         if(is_numeric($answerId)){
-            self::show_delete_answer($questionId, $answerId);  
+            self::show_delete_answer($questionId, $answerId, $user);  
         }else{
-            self::show_delete_question($questionId);
+            self::show_delete_question($questionId, $user);
         }
     }
 
-    private function show_delete_question($questionId){
-        if(self::get_user_or_redirect()->getUserId() === Question::get_question($questionId)->getAuthorId()){
-            (new View("delete"))->show(array("postId"=> $questionId));
+    private function show_delete_question($questionId, $user){
+        if($user->getUserId() === Question::get_question($questionId)->getAuthorId()){
+            (new View("delete"))->show(array("postId"=>$questionId, "user"=>$user));
         } 
     }
     
-    private function show_delete_answer($questionId, $answerId){
-        if(self::get_user_or_redirect()->getUserId() === Answer::get_answer($answerId)->getAuthorId()){
-            (new View("delete"))->show(array("postId"=> $questionId, "answerId"=> $answerId));
+    private function show_delete_answer($questionId, $answerId, $user){
+        if($user->getUserId() === Answer::get_answer($answerId)->getAuthorId()){
+            (new View("delete"))->show(array("postId"=>$questionId, "answerId"=>$answerId, "user"=>$user));
         }
     }  
 
 
-    private function delete_questionOrAnswer($questionId, $answerId){
+    private function delete_questionOrAnswer($questionId, $answerId, $user){
         if(is_numeric($answerId)){
-            self::delete_answer($questionId, $answerId);
+            self::delete_answer($questionId, $answerId, $user);
         }else{
-            self::delete_question($questionId);        
+            self::delete_question($questionId, $user);        
         }
     }
     
-    private function delete_question($questionId){
+    private function delete_question($questionId, $user){
         $question = new Question($questionId, null, null, null, null, null, null, null, null, null, null);
-        if(self::get_user_or_redirect()->getUserId() === Question::get_question($questionId)->getAuthorId() 
-                                                     && $question->delete($questionId)){
+        if($user->getUserId() === Question::get_question($questionId)->getAuthorId() && $question->delete($questionId)){
             self::redirect();    
         }
     }
 
-    private function delete_answer($questionId, $answerId){
+    private function delete_answer($questionId, $answerId, $user){
         $answer = new Answer($answerId, null, $questionId, null, null, null, null);
-        if(self::get_user_or_redirect()->getUserId() === Answer::get_answer($answerId)->getAuthorId() 
-                                                    && $answer->delete($questionId, $answerId)){
+        if($user->getUserId() === Answer::get_answer($answerId)->getAuthorId() && $answer->delete($questionId, $answerId)){
             self::redirect("post", "show", $questionId);    
         }
     }
                 
     public function answer(){
-        if($this->user_logged() && isset($_GET['param1']) && isset($_POST['body'])){
+        $user = self::get_user_or_redirect();
+        if(isset($_GET['param1']) && isset($_POST['body'])){
             $parentId = $_GET['param1'];
             $body = $_POST['body'];
-            $authorId = self::get_user_or_redirect()->getUserId();
+            $authorId = $user->getUserId();
             $answer = new Answer($body, $authorId, $parentId, null, null, null, null);
             $error = Answer::validate($answer);
             if(count($error) == 0){
                 $answer->add_answer();
                 self::redirect("post", "show", $parentId);
             }else{
-                (new View("show"))->show(array("post"=> Question::get_question($parentId), "error" => $error)); 
+                (new View("show"))->show(array("post"=> Question::get_question($parentId), "user"=> $user, "error" => $error)); 
             }           
         }else{
             self::redirect();
@@ -250,10 +272,11 @@ class ControllerPost extends Controller{
     }
 
     public function accept_answer(){
+        $user = self::get_user_or_redirect();
         if(isset($_GET['param1']) && isset($_GET['param2'])){
             $postId = $_GET['param1'];
             $answerId = $_GET['param2'];
-            if(self::get_user_or_redirect()->getUserId() === Question::get_question($postId)->getAuthorId()){
+            if($user->getUserId() === Question::get_question($postId)->getAuthorId()){
                 if(Question::accept_answer($postId, $answerId)){
                     self::redirect("post", "show", $postId);    
                 }
@@ -263,11 +286,11 @@ class ControllerPost extends Controller{
         }   
     }
 
-    // +/- mais pas juste !
     public function delete_accepted_answer(){
+        $user = self::get_user_or_redirect();
         if(isset($_GET['param1']) && isset($_POST['delete_acceptation'])){
             $postId = $_GET['param1'];
-            if(self::get_user_or_redirect()->getUserId() === Question::get_question($postId)->getAuthorId()){
+            if($user->getUserId() === Question::get_question($postId)->getAuthorId()){
                 if(Question::delete_accepted_answer($postId)){
                     self::redirect("post", "show", $postId);
                 }
@@ -276,8 +299,6 @@ class ControllerPost extends Controller{
             self::redirect();
         }   
     }
-
-    
 }
 
 ?>
